refactor(cli): split api verb registry by domain
Move the inline VERBS array out of verbs.ts (500 lines, zero headroom) into
domain files: verbs-read.ts, verbs-create.ts, verbs-drive.ts, verbs-edit.ts,
verbs-lifecycle.ts, verbs-worktree.ts, verbs-feedback.ts, plus task-statuses.ts.
verbs.ts now only holds registry metadata, the schema handler, and the concat.
No verb schema, help text, or runtime behavior changed.
